Make ConfirmStripeIntentParamsFactory a sealed class.
